# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Wallboard Backend API
  description: |
    Backend API für die Wallboard/Dashboard-Anwendung mit Unterstützung für:
    - Dashboard- und Widget-Verwaltung
    - Google Calendar Integration  
    - Spotify Integration
    - SmartThings IoT Integration
    - Real-time Updates über Socket.IO
  version: 1.0.0
  contact:
    name: API Support
    url: https://screen-board.com
  license:
    name: Proprietary
servers:
  - url: https://api.screen-board.com/api
    description: Produktionsserver
  - url: http://localhost:5000/api
    description: Entwicklungsserver
security:
  - bearerAuth: []
...
paths:
...
  /dashboard/{dashboardId}:
    get:
      tags:
        - Dashboard
      summary: Dashboard abrufen
      description: Ruft ein spezifisches Dashboard mit allen zugehörigen Widgets und Layout ab
      operationId: getDashboard
      parameters:
        - name: dashboardId
          in: path
          required: true
          description: Eindeutige Dashboard-ID
          schema:
            type: string
      responses:
        '200':
          description: Dashboard erfolgreich abgerufen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Dashboard nicht gefunden
    
    put:
      tags:
        - Dashboard
      summary: Dashboard aktualisieren
      description: Aktualisiert ein bestehendes Dashboard
      operationId: updateDashboard
      parameters:
        - name: dashboardId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DashboardUpdate'
      responses:
        '200':
          description: Dashboard erfolgreich aktualisiert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Dashboard nicht gefunden
    
    delete:
      tags:
        - Dashboard
      summary: Dashboard löschen
      description: Löscht ein Dashboard und alle zugehörigen Widgets
      operationId: deleteDashboard
      parameters:
        - name: dashboardId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Dashboard erfolgreich gelöscht
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Dashboard nicht gefunden
  /dashboardList:
    get:
      tags:
        - Dashboard
      summary: Liste aller Dashboards
      description: Ruft alle Dashboards des authentifizierten Benutzers ab
      operationId: getDashboardList
      responses:
        '200':
          description: Liste der Dashboards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dashboard'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /dashboard:
    post:
      tags:
        - Dashboard
      summary: Neues Dashboard erstellen
      description: Erstellt ein neues Dashboard für den authentifizierten Benutzer
      operationId: createDashboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DashboardCreate'
      responses:
        '201':
          description: Dashboard erfolgreich erstellt
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '400':
          description: Ungültige Eingabedaten

  # Widget Endpoints
...
  /dashboard/{dashboardId}/layout:
    get:
      tags:
        - Layout
      summary: Layout abrufen
      description: Ruft die Layout-Konfiguration für ein Dashboard ab
      operationId: getLayout
      parameters:
        - name: dashboardId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Layout erfolgreich abgerufen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Layout'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    
    put:
      tags:
        - Layout
      summary: Layout aktualisieren
      description: Aktualisiert die Layout-Konfiguration
      operationId: updateLayout
      parameters:
        - name: dashboardId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LayoutUpdate'
      responses:
        '200':
          description: Layout erfolgreich aktualisiert
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Layout'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # User Endpoints
...
  /smartthings/{dashboardId}/devices/{deviceId}/command:
    post:
      tags:
        - SmartThings
      summary: Gerätebefehl ausführen
      description: Führt einen Befehl auf einem SmartThings-Gerät aus
      operationId: executeDeviceCommand
      parameters:
        - name: dashboardId
          in: path
          required: true
          schema:
            type: string
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - command
                - capability
              properties:
                command:
                  type: string
                  example: "on"
                capability:
                  type: string
                  example: "switch"
                arguments:
                  type: array
                  items: {}
      responses:
        '200':
          description: Befehl erfolgreich ausgeführt
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /smartthings/{dashboardId}/devices/{deviceId}/status:
    get:
      tags:
        - SmartThings
      summary: Gerätestatus abrufen
      description: Ruft den aktuellen Status eines SmartThings-Geräts ab
      operationId: getDeviceStatus
      parameters:
        - name: dashboardId
          in: path
          required: true
          schema:
            type: string
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Gerätestatus
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmartThingsDeviceStatus'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Auth0 JWT Token. Alle API-Endpunkte (außer OAuth-Callbacks und Webhooks) 
        erfordern einen gültigen JWT-Token im Authorization-Header.
  schemas:
    Dashboard:
      type: object
      properties:
        _id:
          type: string
          description: Eindeutige Dashboard-ID
        userId:
          type: string
          description: Benutzer-ID des Eigentümers
        name:
          type: string
          description: Name des Dashboards
        description:
          type: string
          description: Beschreibung des Dashboards
        widgets:
          type: array
          items:
            $ref: '#/components/schemas/Widget'
        layout:
          $ref: '#/components/schemas/Layout'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - _id
        - userId
        - name
    DashboardCreate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
      required:
        - name
    DashboardUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    Widget:
      type: object
      properties:
        _id:
          type: string
        dashboardId:
          type: string
        type:
          type: string
          enum:
            - datetime
            - weather
            - music
            - events
            - notes
            - qrcodes
            - remarks
            - iot
            - news
        config:
          type: object
          description: Widget-spezifische Konfiguration
        position:
          type: object
          properties:
            x:
              type: integer
            y:
              type: integer
            w:
              type: integer
            h:
              type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - _id
        - dashboardId
        - type
...
    Layout:
      type: object
      properties:
        _id:
          type: string
        dashboardId:
          type: string
        cols:
          type: integer
          description: Anzahl der Spalten
        rowHeight:
          type: integer
          description: Höhe einer Zeile in Pixeln
        layouts:
          type: object
          description: Responsive Layout-Konfiguration
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LayoutUpdate:
      type: object
      properties:
        cols:
          type: integer
        rowHeight:
          type: integer
        layouts:
          type: object
...
    SmartThingsDeviceStatus:
      type: object
      properties:
        deviceId:
          type: string
        components:
          type: object
          additionalProperties:
            type: object
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
      required:
        - error
        - message
  responses:
    UnauthorizedError:
      description: Authentifizierung fehlgeschlagen oder Token ungültig
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid token"
            message: "Unauthorized"